requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
blis==1.3.2
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
# """ if en_core_web_sm fails, try downloading and installing manually via: python -m spacy download en_core_web_sm """
fastapi==0.121.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
narwhals==2.11.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
preshed==3.0.11
protobuf==6.33.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.49.3
streamlit==1.51.0
tenacity==9.1.2
thinc==8.3.9
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
